+++ backend/parsers/windmill-parser-py/src/lib.rs
@@ -15,45 +15,19 @@ use windmill_parser::{json_to_typ, Arg, MainArgSignature, ObjectType, Typ};
 
 use rustpython_parser::{
     ast::{
-        Constant, Expr, ExprAttribute, ExprCall, ExprConstant, ExprDict, ExprList, ExprName,
-        Stmt, StmtClassDef, StmtFunctionDef, Suite,
+        Constant, Expr, ExprConstant, ExprDict, ExprList, ExprName,
+        Stmt, StmtFunctionDef, Suite,
     },
     Parse,
 };
 
 pub mod asset_parser;
+pub mod pydantic_parser;
+
 pub use asset_parser::parse_assets;
+use pydantic_parser::ModuleGuard;
 
 const FUNCTION_CALL: &str = "<function call>";
-
-// Thread-local storage for AST module to enable Pydantic/dataclass detection
-// Uses a STACK to support recursive calls without race conditions
-use std::cell::RefCell;
-
-thread_local! {
-    static MODULE_STACK: RefCell<Vec<Vec<Stmt>>> = RefCell::new(Vec::new());
-}
-
-/// Push a module onto the stack
-fn set_current_module(module: Vec<Stmt>) {
-    MODULE_STACK.with(|stack| {
-        stack.borrow_mut().push(module);
-    });
-}
-
-/// Get a clone of the current (top of stack) module
-fn get_current_module() -> Option<Vec<Stmt>> {
-    MODULE_STACK.with(|stack| {
-        stack.borrow().last().cloned()
-    })
-}
-
-/// Pop the current module from the stack
-fn clear_current_module() {
-    MODULE_STACK.with(|stack| {
-        stack.borrow_mut().pop();
-    });
-}
 
 fn filter_non_main(code: &str, main_name: &str) -> String {
     let def_main = format!("def {}(", main_name);
@@ -88,22 +62,6 @@ fn filter_non_main(code: &str, main_name: &str) -> String {
 
     filtered_code.push_str(": return");
     return filtered_code;
-}
-
-/// RAII guard to ensure thread-local module is always cleaned up
-struct ModuleGuard;
-
-impl ModuleGuard {
-    fn new(module: Vec<Stmt>) -> Self {
-        set_current_module(module);
-        Self
-    }
-}
-
-impl Drop for ModuleGuard {
-    fn drop(&mut self) {
-        clear_current_module();
-    }
 }
 
 /// skip_params is a micro optimization for when we just want to find the main
@@ -297,8 +255,8 @@ fn parse_typ(id: &str) -> Typ {
         }
         _ => {
             // Check if it's a Pydantic model or dataclass
-            if let Some(module) = get_current_module() {
-                if let Some(object_type) = detect_model_type(id, &module) {
+            if let Some(module) = pydantic_parser::get_current_module() {
+                if let Some(object_type) = pydantic_parser::detect_model_type(id, &module) {
                     return Typ::Object(object_type);
                 }
             }
@@ -316,300 +274,6 @@ fn map_resource_name(x: &str) -> String {
     }
 }
 
-// ============================================================================
-// Pydantic/Dataclass Detection Functions
-// ============================================================================
-
-/// Maximum number of fields allowed in a Pydantic model or dataclass
-/// This prevents malicious code from defining models with thousands of fields
-const MAX_MODEL_FIELDS: usize = 200;
-
-/// Detects if a class name refers to a Pydantic model or dataclass
-/// Returns ObjectType with parsed fields if detected, None otherwise
-fn detect_model_type(class_name: &str, module: &[Stmt]) -> Option<ObjectType> {
-    // Find class definition in module
-    for stmt in module {
-        if let Stmt::ClassDef(class_def) = stmt {
-            if class_def.name.as_str() == class_name {
-                // Check if it's a Pydantic model (inherits from BaseModel)
-                if is_pydantic_base(&class_def.bases) {
-                    return parse_pydantic_model(class_def);
-                }
-
-                // Check if it's a dataclass (has @dataclass decorator)
-                if has_dataclass_decorator(&class_def.decorator_list) {
-                    return parse_dataclass_fields(&class_def.body, class_def.name.as_str());
-                }
-
-                // Found class but it's neither Pydantic nor dataclass
-                return None;
-            }
-        }
-    }
-
-    // Class not found in module
-    None
-}
-
-/// Checks if a class inherits from BaseModel or pydantic.BaseModel
-fn is_pydantic_base(bases: &[Expr]) -> bool {
-    for base in bases {
-        match base {
-            // Match: class User(BaseModel)
-            Expr::Name(ExprName { id, .. }) => {
-                if id.as_str() == "BaseModel" {
-                    return true;
-                }
-            }
-
-            // Match: class User(pydantic.BaseModel)
-            Expr::Attribute(ExprAttribute { attr, value, .. }) => {
-                if attr.as_str() == "BaseModel" {
-                    if let Expr::Name(ExprName { id, .. }) = value.as_ref() {
-                        if id.as_str() == "pydantic" {
-                            return true;
-                        }
-                    }
-                }
-            }
-
-            _ => {}
-        }
-    }
-
-    false
-}
-
-/// Checks if a class has @dataclass decorator
-fn has_dataclass_decorator(decorators: &[Expr]) -> bool {
-    for decorator in decorators {
-        match decorator {
-            // Match: @dataclass
-            Expr::Name(ExprName { id, .. }) => {
-                if id.as_str() == "dataclass" {
-                    return true;
-                }
-            }
-
-            // Match: @dataclasses.dataclass
-            Expr::Attribute(ExprAttribute { attr, value, .. }) => {
-                if attr.as_str() == "dataclass" {
-                    if let Expr::Name(ExprName { id, .. }) = value.as_ref() {
-                        if id.as_str() == "dataclasses" {
-                            return true;
-                        }
-                    }
-                }
-            }
-
-            // Match: @dataclass() or @dataclass(frozen=True)
-            Expr::Call(ExprCall { func, .. }) => {
-                if let Expr::Name(ExprName { id, .. }) = func.as_ref() {
-                    if id.as_str() == "dataclass" {
-                        return true;
-                    }
-                }
-
-                // Also check for @dataclasses.dataclass(...)
-                if let Expr::Attribute(ExprAttribute { attr, value, .. }) = func.as_ref() {
-                    if attr.as_str() == "dataclass" {
-                        if let Expr::Name(ExprName { id, .. }) = value.as_ref() {
-                            if id.as_str() == "dataclasses" {
-                                return true;
-                            }
-                        }
-                    }
-                }
-            }
-
-            _ => {}
-        }
-    }
-
-    false
-}
-
-/// Parses Pydantic model fields from class body
-fn parse_pydantic_model(class_def: &StmtClassDef) -> Option<ObjectType> {
-    let mut properties = Vec::new();
-
-    for stmt in &class_def.body {
-        // Extract annotated assignments: field_name: field_type
-        if let Stmt::AnnAssign(ann_assign) = stmt {
-            if let Expr::Name(ExprName { id: field_name, .. }) = ann_assign.target.as_ref() {
-                // Security: Enforce field count limit
-                if properties.len() >= MAX_MODEL_FIELDS {
-                    // Log warning and return truncated model
-                    eprintln!(
-                        "Warning: Model '{}' exceeds maximum field count ({}). Truncating to {} fields.",
-                        class_def.name, MAX_MODEL_FIELDS, MAX_MODEL_FIELDS
-                    );
-                    break;
-                }
-
-                // Extract field type from annotation
-                let field_type = extract_field_type(&ann_assign.annotation, 0);
-
-                properties.push(windmill_parser::ObjectProperty {
-                    key: field_name.to_string(),
-                    typ: Box::new(field_type),
-                });
-            }
-        }
-
-        // Skip other statements (methods, class variables without annotations, etc.)
-    }
-
-    if properties.is_empty() {
-        // Empty model - return object with no properties
-        return Some(ObjectType {
-            name: Some(class_def.name.to_string()),
-            props: None,
-        });
-    }
-
-    Some(ObjectType {
-        name: Some(class_def.name.to_string()),
-        props: Some(properties),
-    })
-}
-
-/// Parses dataclass fields from class body (same as Pydantic)
-fn parse_dataclass_fields(body: &[Stmt], class_name: &str) -> Option<ObjectType> {
-    let mut properties = Vec::new();
-
-    for stmt in body {
-        // Extract annotated assignments: field_name: field_type
-        if let Stmt::AnnAssign(ann_assign) = stmt {
-            if let Expr::Name(ExprName { id: field_name, .. }) = ann_assign.target.as_ref() {
-                // Security: Enforce field count limit
-                if properties.len() >= MAX_MODEL_FIELDS {
-                    // Log warning and return truncated model
-                    eprintln!(
-                        "Warning: Dataclass '{}' exceeds maximum field count ({}). Truncating to {} fields.",
-                        class_name, MAX_MODEL_FIELDS, MAX_MODEL_FIELDS
-                    );
-                    break;
-                }
-
-                // Extract field type from annotation
-                let field_type = extract_field_type(&ann_assign.annotation, 0);
-
-                properties.push(windmill_parser::ObjectProperty {
-                    key: field_name.to_string(),
-                    typ: Box::new(field_type),
-                });
-            }
-        }
-    }
-
-    if properties.is_empty() {
-        // Empty dataclass - return object with no properties
-        return Some(ObjectType {
-            name: Some(class_name.to_string()),
-            props: None,
-        });
-    }
-
-    Some(ObjectType {
-        name: Some(class_name.to_string()),
-        props: Some(properties),
-    })
-}
-
-/// Extracts Windmill Typ from Python type annotation (RECURSIVE)
-/// depth: Current recursion depth (prevents infinite recursion)
-fn extract_field_type(annotation: &Expr, depth: u8) -> Typ {
-    const MAX_RECURSION_DEPTH: u8 = 10;
-
-    // Prevent infinite recursion (max depth: 10)
-    if depth >= MAX_RECURSION_DEPTH {
-        eprintln!(
-            "Warning: Type annotation recursion limit ({}) reached. Returning Unknown type.",
-            MAX_RECURSION_DEPTH
-        );
-        return Typ::Unknown;
-    }
-
-    match annotation {
-        // Simple types: str, int, bool, float, bytes
-        Expr::Name(ExprName { id, .. }) => match id.as_str() {
-            "str" => Typ::Str(None),
-            "int" => Typ::Int,
-            "float" => Typ::Float,
-            "bool" => Typ::Bool,
-            "bytes" => Typ::Bytes,
-            "datetime" => Typ::Datetime,
-            // Custom class - check if it's a model
-            custom_type => {
-                // Try to detect if this is a nested Pydantic model
-                if let Some(module) = get_current_module() {
-                    if let Some(object_type) = detect_model_type(custom_type, &module) {
-                        return Typ::Object(object_type);
-                    }
-                }
-                // Unknown type - return Unknown instead of Resource
-                // (Resource is for runtime-resolvable resources)
-                Typ::Unknown
-            }
-        },
-
-        // Generic types: List[T], Optional[T], Dict[K, V]
-        Expr::Subscript(subscript) => {
-            if let Expr::Name(ExprName { id, .. }) = subscript.value.as_ref() {
-                match id.as_str() {
-                    // List[T]
-                    "List" | "list" => {
-                        let inner_type = extract_field_type(&subscript.slice, depth + 1);
-                        Typ::List(Box::new(inner_type))
-                    }
-
-                    // Optional[T] - unwrap to T
-                    "Optional" => extract_field_type(&subscript.slice, depth + 1),
-
-                    // Dict[K, V] - return generic Object
-                    "Dict" | "dict" => Typ::Object(ObjectType::new(None, Some(vec![]))),
-
-                    // Set[T], Tuple[T], etc. - not supported in MVP
-                    _ => Typ::Unknown,
-                }
-            } else {
-                Typ::Unknown
-            }
-        }
-
-        // Union types: str | int (Python 3.10+) or Union[str, int]
-        // MVP: Not supported, return Unknown
-        Expr::BinOp(_) => {
-            eprintln!(
-                "Warning: Union types (e.g., str | int) are not yet supported. \
-                 Field will be treated as Unknown type."
-            );
-            Typ::Unknown
-        }
-
-        // String annotations: "ForwardRef" (forward references)
-        // MVP: Not supported, return Unknown
-        Expr::Constant(ExprConstant {
-            value: Constant::Str(s),
-            ..
-        }) => {
-            eprintln!(
-                "Warning: Forward references (e.g., '{}') are not yet supported. \
-                 Field will be treated as Unknown type.",
-                s
-            );
-            Typ::Unknown
-        }
-
-        // All other annotations
-        _ => Typ::Unknown,
-    }
-}
-
-// ============================================================================
-// End of Pydantic/Dataclass Detection Functions
-// ============================================================================
 
 fn to_value<R>(et: &Expr<R>) -> Option<serde_json::Value> {
     match et {
@@ -1110,424 +774,6 @@ def main(a: str, b: Optional[str], c: str | None): return
                 has_preprocessor: Some(false)
             }
         );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_basic_model() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-
-class User(BaseModel):
-    name: str
-    age: int
-    email: str
-
-def main(user: User):
-    return f'Hello {user.name}'
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        // Check that user parameter is detected as Object type
-        assert_eq!(result.args.len(), 1);
-        assert_eq!(result.args[0].name, "user");
-
-        // Verify it's an Object type with correct model name
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("User".to_string()));
-                assert!(obj.props.is_some());
-
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 3);
-
-                // Verify field names and types
-                assert_eq!(props[0].key, "name");
-                assert_eq!(*props[0].typ, Typ::Str(None));
-
-                assert_eq!(props[1].key, "age");
-                assert_eq!(*props[1].typ, Typ::Int);
-
-                assert_eq!(props[2].key, "email");
-                assert_eq!(*props[2].typ, Typ::Str(None));
-            }
-            _ => panic!("Expected Typ::Object for Pydantic model"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_python_dataclass() -> anyhow::Result<()> {
-        let code = "
-from dataclasses import dataclass
-
-@dataclass
-class Config:
-    host: str
-    port: int
-    debug: bool
-
-def main(config: Config):
-    return config.host
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        // Check that config parameter is detected as Object type
-        assert_eq!(result.args.len(), 1);
-        assert_eq!(result.args[0].name, "config");
-
-        // Verify it's an Object type with correct class name
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("Config".to_string()));
-                assert!(obj.props.is_some());
-
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 3);
-
-                // Verify field names and types
-                assert_eq!(props[0].key, "host");
-                assert_eq!(*props[0].typ, Typ::Str(None));
-
-                assert_eq!(props[1].key, "port");
-                assert_eq!(*props[1].typ, Typ::Int);
-
-                assert_eq!(props[2].key, "debug");
-                assert_eq!(*props[2].typ, Typ::Bool);
-            }
-            _ => panic!("Expected Typ::Object for dataclass"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_nested_model() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-
-class Address(BaseModel):
-    street: str
-    city: str
-
-class Person(BaseModel):
-    name: str
-    address: Address
-
-def main(person: Person):
-    return person.name
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        // Check that person parameter is detected as Object type
-        assert_eq!(result.args.len(), 1);
-        assert_eq!(result.args[0].name, "person");
-
-        // Verify it's an Object type with nested model
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("Person".to_string()));
-                assert!(obj.props.is_some());
-
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                // Verify name field
-                assert_eq!(props[0].key, "name");
-                assert_eq!(*props[0].typ, Typ::Str(None));
-
-                // Verify address field is a nested Object
-                assert_eq!(props[1].key, "address");
-                match props[1].typ.as_ref() {
-                    Typ::Object(nested_obj) => {
-                        assert_eq!(nested_obj.name, Some("Address".to_string()));
-                        assert!(nested_obj.props.is_some());
-
-                        let nested_props = nested_obj.props.as_ref().unwrap();
-                        assert_eq!(nested_props.len(), 2);
-                        assert_eq!(nested_props[0].key, "street");
-                        assert_eq!(nested_props[1].key, "city");
-                    }
-                    _ => panic!("Expected nested Typ::Object for Address"),
-                }
-            }
-            _ => panic!("Expected Typ::Object for Person model"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_empty_model() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-
-class EmptyModel(BaseModel):
-    pass
-
-def main(model: EmptyModel):
-    return 'ok'
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("EmptyModel".to_string()));
-                assert!(obj.props.is_none());
-            }
-            _ => panic!("Expected Typ::Object for empty model"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_list_field() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-from typing import List
-
-class TodoList(BaseModel):
-    items: List[str]
-    count: int
-
-def main(todos: TodoList):
-    return todos.count
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("TodoList".to_string()));
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                // Verify List[str] type
-                assert_eq!(props[0].key, "items");
-                match props[0].typ.as_ref() {
-                    Typ::List(inner) => {
-                        assert_eq!(**inner, Typ::Str(None));
-                    }
-                    _ => panic!("Expected Typ::List for items field"),
-                }
-
-                assert_eq!(props[1].key, "count");
-                assert_eq!(*props[1].typ, Typ::Int);
-            }
-            _ => panic!("Expected Typ::Object"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_optional_field() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-from typing import Optional
-
-class User(BaseModel):
-    name: str
-    nickname: Optional[str]
-
-def main(user: User):
-    return user.name
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                assert_eq!(props[0].key, "name");
-                assert_eq!(*props[0].typ, Typ::Str(None));
-
-                // Optional[str] should unwrap to str
-                assert_eq!(props[1].key, "nickname");
-                assert_eq!(*props[1].typ, Typ::Str(None));
-            }
-            _ => panic!("Expected Typ::Object"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_dataclass_with_decorator_args() -> anyhow::Result<()> {
-        let code = "
-from dataclasses import dataclass
-
-@dataclass(frozen=True)
-class ImmutableConfig:
-    setting: str
-    value: int
-
-def main(config: ImmutableConfig):
-    return config.setting
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("ImmutableConfig".to_string()));
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-            }
-            _ => panic!("Expected Typ::Object for dataclass"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_pydantic_dict_field() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-from typing import Dict
-
-class Config(BaseModel):
-    settings: Dict[str, str]
-    name: str
-
-def main(config: Config):
-    return config.name
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                // Dict should return generic Object
-                assert_eq!(props[0].key, "settings");
-                match props[0].typ.as_ref() {
-                    Typ::Object(_) => {} // Generic object for Dict
-                    _ => panic!("Expected Typ::Object for Dict field"),
-                }
-            }
-            _ => panic!("Expected Typ::Object"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_non_model_class_treated_as_resource() -> anyhow::Result<()> {
-        let code = "
-class RegularClass:
-    def __init__(self, value):
-        self.value = value
-
-def main(obj: RegularClass):
-    return 'ok'
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        // Regular classes (non-Pydantic/dataclass) should be treated as Resource
-        assert_eq!(result.args[0].typ, Typ::Resource("RegularClass".to_string()));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_invalid_syntax_fallback() -> anyhow::Result<()> {
-        // Code with syntax errors - should still not crash
-        let code = "
-from pydantic import BaseModel
-
-class User(BaseModel:  # Missing closing paren
-    name: str
-
-def main(user: User):
-    return 'ok'
-";
-        // Should not panic, even with invalid syntax
-        let result = parse_python_signature(code, None, false);
-
-        // Either succeeds with Unknown types or fails gracefully
-        assert!(result.is_ok() || result.is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_datetime_type() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-from datetime import datetime
-
-class Event(BaseModel):
-    name: str
-    created_at: datetime
-
-def main(event: Event):
-    return event.name
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                assert_eq!(props[0].key, "name");
-                assert_eq!(*props[0].typ, Typ::Str(None));
-
-                assert_eq!(props[1].key, "created_at");
-                assert_eq!(*props[1].typ, Typ::Datetime);
-            }
-            _ => panic!("Expected Typ::Object"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_multiple_pydantic_models() -> anyhow::Result<()> {
-        let code = "
-from pydantic import BaseModel
-
-class User(BaseModel):
-    name: str
-
-class Post(BaseModel):
-    title: str
-    author: User
-
-def main(post: Post):
-    return post.title
-";
-        let result = parse_python_signature(code, None, false)?;
-
-        assert_eq!(result.args.len(), 1);
-        match &result.args[0].typ {
-            Typ::Object(obj) => {
-                assert_eq!(obj.name, Some("Post".to_string()));
-                let props = obj.props.as_ref().unwrap();
-                assert_eq!(props.len(), 2);
-
-                // Nested User model
-                assert_eq!(props[1].key, "author");
-                match props[1].typ.as_ref() {
-                    Typ::Object(nested) => {
-                        assert_eq!(nested.name, Some("User".to_string()));
-                    }
-                    _ => panic!("Expected nested Typ::Object for User"),
-                }
-            }
-            _ => panic!("Expected Typ::Object"),
-        }
 
         Ok(())
     }
